Move relevant tests tests from frame/value_tests to serialize/value_tests
